requirements.txt
PyMySQL==0.8.0
aiomysql
bcrypt
tornado
wit
pycrypto
beautifulsoup4
